Add snapshot test
