+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "plugins": ["transform-object-rest-spread"]
+  "plugins": ["transform-object-rest-spread"],
 }
