+++ .eslintrc.js
@@ -2,9 +2,11 @@ module.exports = {
   env: {
     es6: true,
     node: true,
+    "jest/globals": true,
   },
   parserOptions: {
     ecmaVersion: 2018,
   },
   extends: ["eslint:recommended", "prettier"],
+  plugins: ["jest"],
 }
