+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
 /gatsby-node.js
+/convert-google-document-to-json.js
+/convert-json-to-markdown.js
