+++ src/convert-json-to-markdown.js
@@ -1,0 +1,9 @@
+const json2md = require("json2md")
+const YAML = require("yamljs")
+
+module.exports = ({content, metadata}) => {
+  return `---
+${YAML.stringify(metadata)}---
+
+${json2md(content)}`
+}
